Tue, Dec 6, 2022, 12:57 AM +00:00

+++ problem-sets/filter-more/helpers.c
@@ -7,7 +7,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
     //the average RGB values should stay the same for the old and new image
     //to make greyscale image, RGB values should be equal
-    int avg_rgb_old;
+    float avg_rgb_old;
     long temp = 0;
 
     for(int i = 0; i < height; i++)
